+++ examples/overlay_external_memory.c
@@ -115,7 +115,7 @@ create_overlay ()
   gxr_overlay_submit_texture (overlay, gk_texture);
   gint64 end = g_get_monotonic_time ();
   g_print ("Submit frame took %f ms\n",
-           (end - start) / (1000.));
+           (double) (end - start) / (1000.));
 
   gxr_overlay_show (overlay);
 }
@@ -161,7 +161,7 @@ timeout_callback ()
       destroy_overlay ();
       gint64 end = g_get_monotonic_time ();
       g_print ("Destroy overlay took %f ms\n",
-               (end - start) / (1000.));
+               (double) (end - start) / (1000.));
     }
   else
     {
@@ -170,7 +170,7 @@ timeout_callback ()
       g_timeout_add (250, change_callback, NULL);
       gint64 end = g_get_monotonic_time ();
       g_print ("Create overlay took %f ms\n",
-               (end - start) / (1000.));
+               (double) (end - start) / (1000.));
     }
 
   return TRUE;
