+++ tests/test_overlay.c
@@ -57,8 +57,8 @@ test_overlay_pixbuf ()
   g_assert_nonnull (overlay);
 
   gxr_overlay_set_mouse_scale (overlay,
-                               gdk_pixbuf_get_width (pixbuf),
-                               gdk_pixbuf_get_height (pixbuf));
+                               (float) (gdk_pixbuf_get_width (pixbuf)),
+                               (float) (gdk_pixbuf_get_height (pixbuf)));
 
   gxr_overlay_submit_texture (overlay, texture);
